backend/requirements.txt
torch
transformers
flask
pillow
flask-cors